refactored SpkActions
